Modernize dependencies to support Python 3.10-3.13
- Widen Python range to >=3.10,<3.14 and replace exact 2021-era pins
  with minimum bounds (resolves to numpy 2.2, scipy 1.15, astropy 6.1,
  lmfit 1.3, matplotlib 3.10)
- Remove broken poetry_run:streamlit script entry and stale poetry.lock-bak
- Replace removed APIs: np.float -> float, scipy.ndimage.filters ->
  scipy.ndimage, and deprecated np.in1d -> np.isin

+++ gausspyplus/processing/moment_masking.py
@@ -258,7 +258,7 @@ class MomentMask(object):
         )
 
         for i, rms in tqdm(enumerate(results_list)):
-            if not isinstance(rms, np.float):
+            if not isinstance(rms, float):
                 warnings.warn(f"Problems with entry {i} from resulting parallel_processing list, skipping entry")
                 continue
             else:
